Add Continent Data and Corresponding Table
The first file contains code to wrangle the dataset as well as the new dataset(found in the second file) containing information about the continent each country belongs to. The output of the R file produces a table showing average values for certain attributes, grouped by the continent of each country.